apps/api/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic-settings==2.10.1
python-multipart==0.0.20
langgraph==1.1.3
pillow==12.1.1
